// src/app/shared/models/carteraIntercompanias/carteraTerceros.model.ts
export class CarteraTerceros{
    cliente: number;
    compania: string;
    diaS_DE_CREDITO: number;
    diaS_SIN_CARTA_C: number;
    diasVencimiento: number;
    documento: string;
    empresa: string;
    f_CARTACROBRO: any;
    fechA_GUIA: string;
    fechA_VENCIMIENTO: any;
    fechaCierre: string;
    iD_AREA: number;
    id: number;
    idEmpresa: number;
    interCliente: any;
    interIdCliente: any;
    nocarta: string;
    nombre: string;
    observaciones: string;
    p1A7: number;
    p8A15: number;
    p16A30: number
    p31: number;
    poR_VENCER: number;
    poR_VENCER2: number;
    recogeR_EN: string;
    referencia: string;
    remision: string;
    saldO_SIN_CCOBRO: number;
    serie: string;
    totaL_INTEGRADO: number;
    totaL_SALDO: number;
    totaL_VENCIDOS: number;
    v1A15: number;
    v16A30: number;
    v31A45: number;
    v46A60: number;
    v61A90: number;
    v91A120: number;
    v120: number;
}